Input: atmel_mxt_ts - Bootloader addresses for mXT1664/mXT1188S
The mXT1664S family chips use different mappings for bootloader addresses.
This means that we must look at the family ID to determine which address to
use. There is an additional complication: when we probe and we don't know the
family ID yet, we need to try both possible addresses to find the bootloader.

Signed-off-by: Nick Dyer <nick.dyer@itdev.co.uk>