Input: atmel_mxt_ts - verify Information Block checksum on probe
By reading the information block and the object table into a contiguous
region of memory, we can verify the checksum at probe time. This means we
verify that we are indeed talking to a chip that supports object protocol
correctly. We also detect I2C comms problems much earlier, resulting in
easier diagnosis.

Signed-off-by: Nick Dyer <nick.dyer@itdev.co.uk>
Acked-by: Benson Leung <bleung@chromium.org>